update RSG files with no namespace
