--- src/typings/RsgExample.d.ts
@@ -1,20 +1,0 @@
-declare namespace Rsg {
-	interface MarkdownExample {
-		type: 'markdown';
-		content: string;
-		settings?: Record<string, any>;
-	}
-
-	interface CodeExample {
-		type: 'code';
-		content: string;
-		lang?: string | null;
-		settings?: Record<string, any>;
-	}
-
-	interface RuntimeCodeExample extends CodeExample {
-		evalInContext(a: string): () => any;
-	}
-
-	type Example = RuntimeCodeExample | MarkdownExample;
-}
