+++ src/typings/RsgExample.ts
@@ -1,0 +1,18 @@
+export interface MarkdownExample {
+	type: 'markdown';
+	content: string;
+	settings?: Record<string, any>;
+}
+
+export interface CodeExample {
+	type: 'code';
+	content: string;
+	lang?: string | null;
+	settings?: Record<string, any>;
+}
+
+export interface RuntimeCodeExample extends CodeExample {
+	evalInContext(a: string): () => any;
+}
+
+export type Example = RuntimeCodeExample | MarkdownExample;
