--- src/typings/RsgRequireItResult.d.ts
@@ -1,10 +1,0 @@
-import { ASTNode } from 'ast-types';
-
-declare global {
-	namespace Rsg {
-		interface RequireItResult {
-			require: string;
-			toAST(): ASTNode;
-		}
-	}
-}
