+++ src/typings/RsgRequireItResult.ts
@@ -1,0 +1,6 @@
+import { ASTNode } from 'ast-types';
+
+export interface RequireItResult {
+	require: string;
+	toAST(): ASTNode;
+}
